refactor: rename documentText module to extractText
The name now describes the module's function (extracting text from the
parse tree) rather than a vague noun.
